backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
openai>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
